chore(config): Began updating config to 0.10.0
Began updating neovim config to 0.10.0, however a few issues still
need to be resolved. clang-format was removed as a direct linter from
nvim-lint, as clangd ships with it. Selene's toml config isn't applying
globally. Shellcheck still causing a few errors. Symbols in statuscol
are not appearing. A few deprecated functions need to be updated to
their 0.10.0-compliant counterparts, namely buf_set and buf_get.
Replaced folke's neodev with folke's lazydev, as it is 0.10.0-compliant.

Signed-off-by: 64815328+Eutropios@users.noreply.github.com